Rename and small refactoring
